add base-Renderer and separate & rename renderableComp
