+++ cocos/2d/assembler/graphics/webgl/index.ts
@@ -23,7 +23,7 @@
  THE SOFTWARE.
 */
 
-import { Renderable2D } from '../../../framework/renderable-2d';
+import { UIRenderer } from '../../../framework/renderable-2d';
 import { IAssemblerManager } from '../../../renderer/base';
 import { Graphics } from '../../../components';
 import { graphicsAssembler as graphics } from './graphics-assembler';
@@ -31,7 +31,7 @@ import { graphicsAssembler as graphics } from './graphics-assembler';
 export { earcut } from './earcut';
 
 const graphicsAssemblerManager: IAssemblerManager = {
-    getAssembler (sprite: Renderable2D) {
+    getAssembler (sprite: UIRenderer) {
         return graphics;
     },
 };
