+++ cocos/2d/assembler/sprite/tiled.ts
@@ -23,14 +23,12 @@
  THE SOFTWARE.
 */
 
-
-
 import { IUV } from '../../assets';
 import { Mat4, Vec3, Color } from '../../../core/math';
 import { IRenderData, RenderData } from '../../renderer/render-data';
 import { IBatcher } from '../../renderer/i-batcher';
 import { Sprite } from '../../components/sprite';
-import { Renderable2D } from '../../framework/renderable-2d';
+import { UIRenderer } from '../../framework/renderable-2d';
 import { IAssembler } from '../../renderer/base';
 import { StaticVBChunk } from '../../renderer/static-vb-accessor';
 
@@ -41,7 +39,7 @@ for (let i = 0; i < 4; i++) {
 const matrix = new Mat4();
 
 export const tiled: IAssembler = {
-    createData (sprite: Renderable2D) {
+    createData (sprite: UIRenderer) {
         return sprite.requestRenderData();
     },
