+++ cocos/2d/components/sprite.ts
@@ -37,7 +37,7 @@ import { Vec2 } from '../../core/math';
 import { ccenum } from '../../core/value-types/enum';
 import { clamp } from '../../core/math/utils';
 import { IBatcher } from '../renderer/i-batcher';
-import { Renderable2D, InstanceMaterialType } from '../framework/renderable-2d';
+import { UIRenderer, InstanceMaterialType } from '../framework/renderable-2d';
 import { PixelFormat } from '../../core/assets/asset-enum';
 import { TextureBase } from '../../core/assets/texture-base';
 import { Material, RenderTexture } from '../../core';
@@ -175,7 +175,7 @@ enum EventType {
 @help('i18n:cc.Sprite')
 @executionOrder(110)
 @menu('2D/Sprite')
-export class Sprite extends Renderable2D {
+export class Sprite extends UIRenderer {
     /**
      * @en
      * The sprite atlas where the sprite is.
