+++ cocos/2d/framework/deprecated.ts
@@ -32,7 +32,7 @@
 import { markAsWarning, removeProperty, replaceProperty } from '../../core/utils';
 import { UIComponent } from './ui-component';
 import { UITransform } from './ui-transform';
-import { Renderable2D } from './renderable-2d';
+import { UIRenderer } from './renderable-2d';
 import { Canvas } from './canvas';
 import { js } from '../../core/utils/js';
 import { legacyCC } from '../../core/global-exports';
@@ -47,7 +47,7 @@ removeProperty(UIComponent.prototype, 'UIComponent', [
     },
 ]);
 
-removeProperty(Renderable2D.prototype, 'Renderable2D.prototype', [
+removeProperty(UIRenderer.prototype, 'Renderable2D.prototype', [
     {
         name: 'srcBlendFactor',
     },
@@ -142,13 +142,13 @@ js.setClassAlias(UITransform, 'cc.UITransformComponent');
  * Alias of [[Renderable2D]]
  * @deprecated Since v1.2
  */
-export { Renderable2D as RenderComponent };
+export { UIRenderer as RenderComponent };
 /**
  * Alias of [[Renderable2D]]
  * @deprecated Since v3.0
  */
-export { Renderable2D as UIRenderable };
-js.setClassAlias(Renderable2D, 'cc.RenderComponent');
+export { UIRenderer as UIRenderable };
+js.setClassAlias(UIRenderer, 'cc.RenderComponent');
 
 /**
  * Alias of [[Canvas]]
@@ -157,3 +157,11 @@ js.setClassAlias(Renderable2D, 'cc.RenderComponent');
 export { Canvas as CanvasComponent };
 legacyCC.CanvasComponent = Canvas;
 js.setClassAlias(Canvas, 'cc.CanvasComponent');
+
+/**
+ * Alias of [[Renderable2D]]
+ * @deprecated Since v3.6
+ */
+export { UIRenderer as Renderable2D };
+legacyCC.internal.Renderable2D = UIRenderer;
+js.setClassAlias(UIRenderer, 'cc.Renderable2D');
