+++ cocos/core/assets/material.ts
@@ -31,7 +31,6 @@
 import { ccclass, serializable, type } from 'cc.decorator';
 import { Asset } from './asset';
 import { EffectAsset } from './effect-asset';
-import { RenderableComponent } from '../components/renderable-component';
 import { Texture, Type } from '../gfx';
 import { TextureBase } from './texture-base';
 import { legacyCC } from '../global-exports';
@@ -41,6 +40,7 @@ import { Color } from '../math/color';
 import { warnID } from '../platform/debug';
 import { Vec4 } from '../math';
 import { SRGBToLinear } from '../pipeline/pipeline-funcs';
+import { BaseRenderer } from '../components';
 
 const v4_1 = new Vec4();
 
@@ -182,7 +182,7 @@ export class Material extends Asset {
      * @en The owner render component
      * @zh 该材质所归属的渲染组件
      */
-    get owner (): RenderableComponent | null {
+    get owner (): BaseRenderer | null {
         return null;
     }
