+++ cocos/core/components/index.ts
@@ -30,6 +30,7 @@ export { default as MissingScript } from './missing-script';
 export { EventHandler } from './component-event-handler';
 export { Component } from './component';
 export { Camera } from './camera-component';
-export { RenderableComponent } from './renderable-component';
+export { Renderer } from './renderable-component';
+export { BaseRenderer } from './base-renderer';
 /** deprecated */
 export * from './deprecated';
