+++ cocos/core/renderer/core/material-instance.ts
@@ -28,15 +28,16 @@
  * @module material
  */
 
-import { RenderableComponent } from '../../components/renderable-component';
+import { Renderer } from '../../components/renderable-component';
 import { Material } from '../../assets/material';
 import { PassInstance } from './pass-instance';
 import { MacroRecord } from './pass-utils';
 import { PassOverrides } from './pass';
+import { BaseRenderer } from '../../components/base-renderer';
 
 export interface IMaterialInstanceInfo {
     parent: Material;
-    owner?: RenderableComponent;
+    owner?: BaseRenderer;
     subModelIdx?: number;
 }
 
@@ -56,7 +57,7 @@ export class MaterialInstance extends Material {
     protected _passes: PassInstance[] = [];
 
     private _parent: Material;
-    private _owner: RenderableComponent | null;
+    private _owner: BaseRenderer | null;
     private _subModelIdx = 0;
 
     constructor (info: IMaterialInstanceInfo) {
