+++ cocos/core/scene-graph/node-ui-properties.ts
@@ -29,7 +29,7 @@
  */
 
 import { JSB } from 'internal:constants';
-import { Renderable2D } from '../../2d/framework/renderable-2d';
+import { UIRenderer } from '../../2d/framework/renderable-2d';
 import { UITransform } from '../../2d/framework/ui-transform';
 import { warnID } from '../platform/debug';
 import { UIMeshRenderer } from '../../2d';
@@ -61,7 +61,7 @@ export class NodeUIProperties {
     get uiComp () {
         return this._uiComp;
     }
-    set uiComp (comp: UIMeshRenderer | Renderable2D | null) {
+    set uiComp (comp: UIMeshRenderer | UIRenderer | null) {
         if (this._uiComp && comp) {
             warnID(12002);
             return;
@@ -69,7 +69,7 @@ export class NodeUIProperties {
         this._uiComp = comp;
     }
 
-    private _uiComp: UIMeshRenderer | Renderable2D | null = null;
+    private _uiComp: UIMeshRenderer | UIRenderer | null = null;
 
     /**
      * @en The opacity of the UI node for final rendering
